SSO_server/requirements.txt
Flask==3.1.1
Flask-RESTful==0.3.10
Flask-JWT-Extended==4.6.0
mysql-connector-python
python-dotenv
bcrypt
jwt
